Allows delete conversation
